+++ src/main/java/org/navneev/index/distance/ArrayBasedDistanceCalculator.java
@@ -113,7 +113,7 @@ public class ArrayBasedDistanceCalculator implements DistanceCalculator {
      * @throws NullPointerException      if memorySegment is null
      */
     @Override
-    public double calculate(final MemorySegment memorySegment) {
+    public float calculate(final MemorySegment memorySegment) {
         return VectorDistanceCalculationUtils.euclideanDistance(memorySegment, baseVector);
     }
 
@@ -133,7 +133,7 @@ public class ArrayBasedDistanceCalculator implements DistanceCalculator {
      * @throws NullPointerException     if vector2 is null
      */
     @Override
-    public double calculate(float[] vector2) {
+    public float calculate(float[] vector2) {
         return VectorDistanceCalculationUtils.euclideanDistance(baseVector, vector2);
     }
